Calculator: Fix Sqrt property
In the past the Sqrt property returns a double value which can be infinity
while a BigInteger value reaches a certain size.
Therefor the return value of Sqrt is changed from double to Calculator{TNumber}.
It is also necessary to re-implement Sqrt algorithm for the BigIntCalculator.
The algorithm based on Heron's method.

Resolves: No entry