Secret.cs: Fix ToBase64 and ParseBase64 method
The big integer data type doesn't keep the
original byte array. Based on the MSB additional bytes
are added for the sign. To avoid this behavior a mask byte
will be added to the byte array if the base64 string is
decoded. On the other side if a base64 string is encoded,
the mask byte will be removed.

Resolves: #60