Secret.cs: Final fix for bug #60 ("Reconstruction fails at random")
Deserialization of secret sources like byte arrays, strings and numbers
into positive big integer values. Positive big integer values are necessary for
the applied modular arithmetic.

This fix results in a breaking change. Shares and secrets which are generated by v0.6.0 or older can't reconstructed.
A legacy mode is introduced to alleviate this breaking change. Set Secret.LegacyMode.Value to true to enable this mode.

It is recommended to reconstruct the secrets in legacy mode on and to split it again in the legacy mode off.

Resolves: #60